Extract each message from received network buffer stream
There maybe several messages received by the node at one time,
each of them should be split for further process, another case
is that the received buffer is just piece of segment of a message.
all above caces need be handled to get correctly message headers
and content
