+++ node/message.go
@@ -1,7 +1,7 @@
-package net
+package node
 
 import (
-	"log"
+	"fmt"
 	"bytes"
 	"errors"
 	"time"
@@ -121,18 +121,6 @@ type transaction struct {
 	// TBD
 }
 
-// TODO Sample function, shoule be called from ledger module
-func ledgerGetHeader() ([]byte, error) {
-	genesisHeader := "b3181718ef6167105b70920e4a8fbbd0a0a56aacf460d70e10ba6fa1668f1fef"
-
-	h, err := hex.DecodeString(genesisHeader)
-	if err != nil {
-		log.Printf("Decode Header hash error")
-		return nil, err
-	}
-	return h, nil
-}
-
 // Alloc different message stucture
 // @t the message name or type
 // @len the message length only valid for varible length structure
@@ -141,8 +129,11 @@ func ledgerGetHeader() ([]byte, error) {
 // @messager the messager structure
 // @error  error code
 // FixMe fix the ugly multiple return.
-func allocMsg(t string, length uint64) (messager, error) {
+func allocMsg(t string, length int) (messager, error) {
 	switch t {
+	case "msgheader":
+		var msg msgHdr
+		return &msg, nil
 	case "version":
 		var msg version
 		return &msg, nil
@@ -204,8 +195,8 @@ func (hdr *msgHdr) init(cmd string, checksum []byte, length uint32) {
 	copy(hdr.Checksum[:], checksum[:CHECKSUMLEN])
 	hdr.Length = length
 
-	log.Printf("The message payload length is %d", hdr.Length)
-	log.Printf("The message header length is %d", uint32(unsafe.Sizeof(*hdr)))
+	fmt.Printf("The message payload length is %d", hdr.Length)
+	fmt.Printf("The message header length is %d", uint32(unsafe.Sizeof(*hdr)))
 }
 
 
@@ -224,7 +215,7 @@ func newVersion() ([]byte, error) {
 	msg.P.TimeStamp = uint32(time.Now().UTC().UnixNano())
 	msg.P.Port = nodes.node.port
 	msg.P.Nonce = nodes.node.nonce
-	log.Printf("The nonce is 0x%x", msg.P.Nonce)
+	fmt.Printf("The nonce is 0x%x", msg.P.Nonce)
 	msg.P.UserAgent = 0x00
 	// Fixme Get the block height from ledger
 	msg.P.StartHeight = 1
@@ -240,7 +231,7 @@ func newVersion() ([]byte, error) {
 	p := new(bytes.Buffer)
 	err := binary.Write(p, binary.LittleEndian, &(msg.P))
 	if err != nil {
-		log.Println("Binary Write failed at new Msg")
+		fmt.Println("Binary Write failed at new Msg")
 		return nil, err
 	}
 	s := sha256.Sum256(p.Bytes())
@@ -249,16 +240,16 @@ func newVersion() ([]byte, error) {
 	buf := bytes.NewBuffer(s[:4])
 	binary.Read(buf, binary.LittleEndian, &(msg.Hdr.Checksum))
 	msg.Hdr.Length = uint32(len(p.Bytes()))
-	log.Printf("The message payload length is %d", msg.Hdr.Length)
+	fmt.Printf("The message payload length is %d\n", msg.Hdr.Length)
 
 	m, err := msg.serialization()
 	if (err != nil) {
-		log.Println("Error Convert net message ", err.Error())
+		fmt.Println("Error Convert net message ", err.Error())
 		return nil, err
 	}
 
 	str := hex.EncodeToString(m)
-	log.Printf("The message length is %d, %s", len(m), str)
+	fmt.Printf("The message length is %d, %s\n", len(m), str)
 	return m, nil
 }
 
@@ -275,7 +266,7 @@ func newVerack() ([]byte, error) {
 	}
 
 	str := hex.EncodeToString(buf)
-	log.Printf("The message tx verack length is %d, %s", len(buf), str)
+	fmt.Printf("The message tx verack length is %d, %s", len(buf), str)
 
 	return buf, err
 }
@@ -293,9 +284,22 @@ func newGetAddr() ([]byte, error) {
 	}
 
 	str := hex.EncodeToString(buf)
-	log.Printf("The message get addr length is %d, %s", len(buf), str)
+	fmt.Printf("The message get addr length is %d, %s", len(buf), str)
 
 	return buf, err
+}
+
+func magicVerify(magic uint32) bool {
+	if (magic != NETMAGIC) {
+		return false
+	}
+	return true
+}
+
+func payloadLen(buf []byte) int {
+	var h msgHdr
+	h.deserialization(buf)
+	return int(h.Length)
 }
 
 func msgType(buf []byte) (string, error) {
@@ -328,7 +332,7 @@ func newHeadersReq() ([]byte, error) {
 
 	// Fixme correct with the exactly request length
 	h.p.len = 1
-	buf, err := ledgerGetHeader()
+	buf, err := LedgerGetHeader()
 	if (err != nil) {
 		return nil, err
 	}
@@ -337,7 +341,7 @@ func newHeadersReq() ([]byte, error) {
 	p := new(bytes.Buffer)
 	err = binary.Write(p, binary.LittleEndian, &(h.p))
 	if err != nil {
-		log.Println("Binary Write failed at new headersReq")
+		fmt.Println("Binary Write failed at new headersReq")
 		return nil, err
 	}
 
@@ -346,7 +350,7 @@ func newHeadersReq() ([]byte, error) {
 
 	m, err := h.serialization()
 	str := hex.EncodeToString(m)
-	log.Printf("The message length is %d, %s", len(m), str)
+	fmt.Printf("The message length is %d, %s\n", len(m), str)
 	return m, err
 }
 
@@ -392,9 +396,6 @@ func (msg inv) verify(buf []byte) error {
 // most of them are the same
 func (hdr msgHdr) serialization() ([]byte, error) {
 	var buf bytes.Buffer
-
-	log.Printf("The size of messge is %d in serialization",
-		uint32(unsafe.Sizeof(hdr)))
 	err := binary.Write(&buf, binary.LittleEndian, hdr)
 	if err != nil {
 		return nil, err
@@ -404,10 +405,8 @@ func (hdr msgHdr) serialization() ([]byte, error) {
 }
 
 func (msg *msgHdr) deserialization(p []byte) error {
-	log.Printf("The size of messge is %d in deserialization",
-		uint32(unsafe.Sizeof(*msg)))
 
-	buf := bytes.NewBuffer(p)
+	buf := bytes.NewBuffer(p[0 : MSGHDRLEN])
 	err := binary.Read(buf, binary.LittleEndian, msg)
 	return err
 }
@@ -415,7 +414,7 @@ func (msg *msgHdr) deserialization(p []byte) error {
 func (msg version) serialization() ([]byte, error) {
 	var buf bytes.Buffer
 
-	log.Printf("The size of messge is %d in serialization",
+	fmt.Printf("The size of messge is %d in serialization\n",
 		uint32(unsafe.Sizeof(msg)))
 	err := binary.Write(&buf, binary.LittleEndian, msg)
 	if err != nil {
@@ -426,7 +425,7 @@ func (msg version) serialization() ([]byte, error) {
 }
 
 func (msg *version) deserialization(p []byte) error {
-	log.Printf("The size of messge is %d in deserialization",
+	fmt.Printf("The size of messge is %d in deserialization\n",
 		uint32(unsafe.Sizeof(*msg)))
 
 	buf := bytes.NewBuffer(p)
@@ -437,7 +436,7 @@ func (msg *version) deserialization(p []byte) error {
 func (msg headersReq) serialization() ([]byte, error) {
 	var buf bytes.Buffer
 
-	log.Printf("The size of messge is %d in serialization",
+	fmt.Printf("The size of messge is %d in serialization\n",
 		uint32(unsafe.Sizeof(msg)))
 	err := binary.Write(&buf, binary.LittleEndian, msg)
 	if err != nil {
@@ -448,7 +447,7 @@ func (msg headersReq) serialization() ([]byte, error) {
 }
 
 func (msg *headersReq) deserialization(p []byte) error {
-	log.Printf("The size of messge is %d in deserialization",
+	fmt.Printf("The size of messge is %d in deserialization\n",
 		uint32(unsafe.Sizeof(*msg)))
 
 	buf := bytes.NewBuffer(p)
@@ -459,7 +458,7 @@ func (msg *headersReq) deserialization(p []byte) error {
 func (msg blkHeader) serialization() ([]byte, error) {
 	var buf bytes.Buffer
 
-	log.Printf("The size of messge is %d in serialization",
+	fmt.Printf("The size of messge is %d in serialization\n",
 		uint32(unsafe.Sizeof(msg)))
 	err := binary.Write(&buf, binary.LittleEndian, msg)
 	if err != nil {
@@ -470,7 +469,7 @@ func (msg blkHeader) serialization() ([]byte, error) {
 }
 
 func (msg *blkHeader) deserialization(p []byte) error {
-	log.Printf("The size of messge is %d in deserialization",
+	fmt.Printf("The size of messge is %d in deserialization\n",
 		uint32(unsafe.Sizeof(*msg)))
 
 	buf := bytes.NewBuffer(p)
@@ -481,7 +480,7 @@ func (msg *blkHeader) deserialization(p []byte) error {
 func (msg addrReq) serialization() ([]byte, error) {
 	var buf bytes.Buffer
 
-	log.Printf("The size of messge is %d in serialization",
+	fmt.Printf("The size of messge is %d in serialization\n",
 		uint32(unsafe.Sizeof(msg)))
 	err := binary.Write(&buf, binary.LittleEndian, msg)
 	if err != nil {
@@ -492,7 +491,7 @@ func (msg addrReq) serialization() ([]byte, error) {
 }
 
 func (msg *addrReq) deserialization(p []byte) error {
-	log.Printf("The size of messge is %d in deserialization",
+	fmt.Printf("The size of messge is %d in deserialization\n",
 		uint32(unsafe.Sizeof(*msg)))
 
 	buf := bytes.NewBuffer(p)
@@ -503,7 +502,7 @@ func (msg *addrReq) deserialization(p []byte) error {
 func (msg inv) serialization() ([]byte, error) {
 	var buf bytes.Buffer
 
-	log.Printf("The size of messge is %d in serialization",
+	fmt.Printf("The size of messge is %d in serialization\n",
 		uint32(unsafe.Sizeof(msg)))
 	err := binary.Write(&buf, binary.LittleEndian, msg)
 	if err != nil {
@@ -514,7 +513,7 @@ func (msg inv) serialization() ([]byte, error) {
 }
 
 func (msg *inv) deserialization(p []byte) error {
-	log.Printf("The size of messge is %d in deserialization",
+	fmt.Printf("The size of messge is %d in deserialization\n",
 		uint32(unsafe.Sizeof(*msg)))
 
 	buf := bytes.NewBuffer(p[0 : MSGHDRLEN])
@@ -551,3 +550,4 @@ func (msg inv) invLen() (uint64, uint8) {
 
 	return val, size
 }
+
